feat(github): add GitHub Actions workflows for build, test, lint, code coverage, and Prisma migration
This commit adds several GitHub Actions workflows to the repository. The
first workflow, "Build Test Lint", is triggered on pull requests to the main
branch and runs the build, test, and lint scripts. The second workflow,
"Code coverage", is also triggered on pull requests to the main branch and
uses the jest-coverage-report-action to generate a code coverage report. The
third workflow, "Prisma migration", is triggered on pushes to the main branch
and runs the Prisma migration deploy command using the DATABASE_URL secret.

Additionally, the Prisma client generator configuration file was updated to
set the "postinstall" option to false.