feat(worker): add support for updating cryptocurrency and exchange rate markets
This commit adds support for updating cryptocurrency and exchange rate markets. The `MarketUpdater` class is introduced with methods to fetch and parse data from the CoinGecko and OpenExchangeRates APIs respectively. The `updateCryptoMarkets` and `updateExchangeRates` functions are added to update the markets for cryptocurrencies and exchange rates respectively. The `markets` function is added to update both markets.